Create diceRoller.dart
It holds the logic behind the rolling of the dice. I have used a StatefulWidget and inside it i have used the "randomizer" function for generating a random number between 1 to 6( randomizer.nextInt(7)+1). It also holds the related settings needed for uploading the image and fitting it in the screen along with other related functionalities.